Move another OAuth2 test from stress.rs to oauth2.rs
