+++ src/oauth2.rs
@@ -398,6 +398,16 @@ mod tests {
     }
 
     #[test]
+    fn test_dc_get_oauth2_url() {
+        let ctx = dummy_context();
+        let addr = "dignifiedquire@gmail.com";
+        let redirect_uri = "chat.delta:/com.b44t.messenger";
+        let res = dc_get_oauth2_url(&ctx.ctx, addr, redirect_uri);
+
+        assert_eq!(res, Some("https://accounts.google.com/o/oauth2/auth?client_id=959970109878%2D4mvtgf6feshskf7695nfln6002mom908%2Eapps%2Egoogleusercontent%2Ecom&redirect_uri=chat%2Edelta%3A%2Fcom%2Eb44t%2Emessenger&response_type=code&scope=https%3A%2F%2Fmail.google.com%2F%20email&access_type=offline".into()));
+    }
+
+    #[test]
     fn test_dc_get_oauth2_token() {
         let ctx = dummy_context();
         let addr = "dignifiedquire@gmail.com";
